World object loot kind of working now
Loot is flagged as "skinning loot" right now, which seems wrong but the
results at the client are correct - whereas corpse loot will be
immediatly rejected by the client. However, other cores DO send corpse
loot there and the client accepts, so something has to be wrong with the
state of objects or something.