Fixed spell effect LearnSpell
This effect didn't work correctly when used with items (recipes), which
executed spell 483. This spell has the LearnSpell effect, but no spell
assigned. Instead, the spell id to learn is written in the items
spellCast list with a special trigger (6).
I also added a new enum for item spell cast trigger types.