Added new navmesh flag
ADTUnwalkable flag is used for terrain portions which players can't walk
on. These are required for multiple purposes, for example creatures
never walk on these either while randomly moving. They only walk on them
when chasing a player (or charging using a spell).
Also, with this, the navigation data can be used for blink.