Fixed wrong creature movement
Creatures seemed to teleport after they reached the last point before
the target point. This was caused by incorrect interpolation due to a
missing point in UnitMover's m_path variable, which does the
interpolation.
I also added a new UNIT_DEBUG_MODE macro which defaults to 0, but can be
set to 1 to enable creature debugging. In short, this will enable
movement logs for all spawned creatures and will only spawn exactly one
timber wolf in northshire, so that the server has less workload and logs
will only be produced by this single creature.