Improved fly behaviour
Flying mounts will be able to fly now.
Creatures won't start attacking players that are flying.
Creatures will reset immediatly if a player starts to fly.
If movement flags start to contain the Flying2 flag, the server will
check for the Fly or ModFlightSpeedMounted aura, and if it can't find
any of these, will kick the player from the server. This should prevent
fly hacks (tested and seemed to work so far).