Client ack codes are handled now
There are still some ToDo's: Speed-acks need to apply the movement speed
only after the ack has been received instead of calculating the speed
using the auras. Also, acks need to be announced and it has to be
checked if a received ack is valid and has been expected. Also, if there
are pending acks, the movement handler may only tolerate a few movement
packets before kicking the player (to fight ack delay to be un-rootable
etc.)