Client Acks will be announced
The server now checks client ack packets
Server kicks players who send acks with unexpected counter values or
unexpected op code values, which ensures that acks have to be sent in
the correct order with the correct data
Acks will now be checked for timeout in order to make it impossible for
players to skip acks entirely.
TODO: Speed is currently applied immediatly at the server, and roots are
checked based on auras instead of move flags. This needs to be adjusted
in order to perform more movement checks based on acks.