Quest changes
A rewarding npc will perform the spell cast now instead of the player
character. If the quest rewarder isn't a GameUnit, the player character
will instead do the spell cast on himself as it was handled until now.
Still not sure if this is the correct behaviour.
Also changed a getLocation() call into getPredictedPosition(time) at
creature combat ai state, so that the creature will always use the
predicted position instead of the last one sent.